Support use of externally created git repositories
This change allows a _pygit2.repository to be constructed around a
struct git_repository (wrapped in a PyCapsule) that has been created
elsewhere. This facility is key to supporting git repositories with
custom storage backends.

The existing api to repository.__init__ is not changed, but extended,
with a git repository path of None indicating that a wrapped struct
git_repository is available in a keyword argument.